+++ app/service/agits/[agitId]/introduce/edit/page.module.css
@@ -1,0 +1,26 @@
+.checkboxLabel {
+  position: relative;
+  display: block;
+  padding: 4px 15px;
+  border-radius: 20px;
+  border: 1px solid #dbdbdb;
+  cursor: pointer;
+  font-weight: bold;
+  font-size: 14px;
+  box-sizing: border-box;
+  transition: color 0.3s ease;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+
+.checkboxInput {
+  display: none;
+}
+
+.checkboxInput:checked + .checkboxLabel {
+  background-color: var(--point-yellow-3);
+  border-color: var(--point-yellow-3);
+  animation: popEffectLight 0.25s cubic-bezier(0.25, 1.3, 0.5, 1);
+}
