+++ app/service/agits/[agitId]/page.module.css
@@ -1,0 +1,3 @@
+.feed_list ul li {
+  width: calc(50% - 5px);
+}
