--- app/service/agits/[id]/page.module.css
@@ -1,3 +1,0 @@
-.feed_list ul li {
-  width: calc(50% - 5px);
-}
