+++ constants/tabMenuList/agits.js
@@ -1,7 +1,7 @@
 export const tabMenuList = [
   { text: '소개', href: `/service/agits/:id/introduce` },
   { text: '홈', href: `/service/agits/:id` },
-  { text: '정모', href: `/service/agits/:id/event` },
+  { text: '정모', href: `/service/agits/:id/meeting` },
   { text: '기록', href: `/service/agits/:id/feed` },
   { text: '아지트 관리', href: `/service/agits/:id/setting` },
 ];
