+++ styles/reset.css
@@ -208,7 +208,8 @@ input[type='password'],
 input[type='tel'],
 input[type='email'],
 input[type='number'],
-input[type='url'] {
+input[type='url'],
+input[type='datetime-local'] {
   text-overflow: ellipsis;
   overflow: hidden;
   display: -webkit-box;
@@ -220,7 +221,8 @@ input[type='text'],
 input[type='password'],
 input[type='tel'],
 input[type='email'],
-input[type='number'] {
+input[type='number'],
+input[type='datetime-local'] {
   display: block;
   width: 100%;
   height: 50px;
@@ -237,6 +239,7 @@ input[type='password'].error,
 input[type='tel'].error,
 input[type='email'].error,
 input[type='number'].error,
+input[type='datetime-local'].error,
 textarea.error {
   border: 1px solid #ea0000;
 }
@@ -246,6 +249,7 @@ input[type='password'].error:focus,
 input[type='tel'].error:focus,
 input[type='email'].error:focus,
 input[type='number'].error:focus,
+input[type='datetime-local'].error:focus,
 textarea.error:focus {
   outline-color: #ea0000;
   box-sizing: border-box;
@@ -256,6 +260,7 @@ input[type='password']::placeholder,
 input[type='tel']::placeholder,
 input[type='email']::placeholder,
 input[type='number']::placeholder,
+input[type='datetime-local']::placeholder,
 textarea::placeholder {
   font-size: 14px;
   font-weight: 500;
@@ -267,6 +272,7 @@ input[type='password']:disabled,
 input[type='tel']:disabled,
 input[type='email']:disabled,
 input[type='number']:disabled,
+input[type='datetime-local']:disabled,
 textarea:disabled {
   background-color: var(--gray-bg);
 }
